% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.r
\name{C.f}
\alias{C.f}
\title{Calculate the incremental carbon footprint of a person attending a TESA activity}
\usage{
C.f(hotel.nights, plane.distance, bustrain.distance, car.distance,
  number.car.sharing, meals)
}
\arguments{
\item{hotel.nights}{number of nights staying at a hotel solo in a room}

\item{plane.distance}{the total flight distance (back and forth) from home to the activity. Set as 0 if you did not take the plane (km)}

\item{bustrain.distance}{the total distance travelled on a bus or train for the activity. Set as 0 if you did not take a bus or train (km)}

\item{car.distance}{the total distance travelled in a car for the activity. Set as 0 if you did not take a car (km)}

\item{number.car.sharing}{the number of people who shared the car to travel to and from the activity}

\item{meals}{the number of meals eaten out during the trip (do not include meals you made yourself)}
}
\description{
This is a rough CO2 calculator. The output is in tonnes CO2 for the trip. It is based on
        information from the carboon footprint calculator. It makes several assumptions: an average meal out
        costs 20 USD. The average hotel cost is 160 USD/night, bustrain carbon cost is an average of several
        modes of wheeled public transport, car travel is based on a standard gasoline car with engine > 2l.
        Meal carbon costs are discounted by 50% because you are not eating at home but because restaurants
        are about half as efficient as home for carbon use mostly owing to food waste (cbc article). Results
        compare pretty well with https://calculator.carbonfootprint.com but they tend to be slightly lower
        primarily because restaurant meal carbon costs are discounted by 50%.

        The carbon footprint is incremental in that it assumes most of the carbon consuming activities that a
        participant does at home (except meals) continue at home while they are away. e.g. they still heat their
        house, feed the dog. The incremental cost might be slightly high but given the result is less than
        the total carbon cost from carbonfootprint, we may have struck a balance. We also consider this
        most useful in comparison between TESA activities and years rather than as an absolute CO2 emission
        measure.
}
\examples{
C.f(hotel.nights=5, plane.distance=1200, bustrain.distance= 0, car.distance=40, number.car.sharing=1, meals=15)
}
\author{
Daniel Duplisea
}
