% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.calcs.r
\name{carbon.footprint.f}
\alias{carbon.footprint.f}
\title{Calculate the carbon footprint for meetings based on transportation, hotel and restaurant meals}
\usage{
carbon.footprint.f(input, Title.name = "Carbon footprint",
  list.out = T)
}
\arguments{
\item{Title.name}{A text string for how you want the plot named}

\item{list.out}{logical, output the list to the console}

\item{input.data}{an object in the workspace with at least columns exactly named:
"activity.type"
"activity.name"
"origin"
"destination"
"origin.country"
"hotel.nights"
"meals"
"bustrain.distance"
"car.distance"
"car.sharing"}
}
\description{
Calculates the carbon emissions footprint for all the activites. It plots the calculation
       as a bar graph with total and per capita values. It also produces a map with all the home cities of
       participants as well as the destination cities which are circled in red. The total carbon emissions
       of all activities combined is shown in the title. A list with calculations is also produced.
}
