% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{carbon.params}
\alias{carbon.params}
\title{Carbon parameters}
\format{A list}
\usage{
data(params)
}
\description{
A list with different carbon emission rates for one unit of travel (km), hotel (nights) and restaurant food (meals)
}
\details{
\itemize{
  \item C.plane carbon emissions per km of plane travel per person
  \item C.car carbon emissions per km of car travel for a car with a motor > 2l
  \item C.bustrain carbon emissions per km of bus or train travel per person
  \item C.hotel carbon emissions per night in an average hotel (160 USD/night)
  \item C.meal carbon emissions per restaurant meal (20 USD/meal), average over breakfast, lunch, supper
  \item C.meal.discount the proportional reduction in restaurant meal carbon to compensate for the fact
        that participants are not eating a meal at home if they are eating one in a restaurant. A CBC article
        cited a value showing that eating a meal in a restaurant emits about 2X the amount of carbon per meal
        than one does at home owing to food waste and other forms of additional carbon consumption associated
        with restaurants
}
}
\keyword{datasets}
