% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.calcs.r
\name{distance.lookup.f}
\alias{distance.lookup.f}
\title{Lookup flying distances between locations}
\usage{
distance.lookup.f(input.data)
}
\arguments{
\item{input.data}{an object in the workspace with at least columns exactly named:
"activity.type"
"activity.name"
"origin"
"destination"
"origin.country"
"hotel.nights"
"meals"
"bustrain.distance"
"car.distance"
"car.sharing"
"flying"}
}
\description{
This looks up flying distances between origin and destination airports. Flying distances are
       taken as direct geographic distances between cities and does not accounting for routing. It has some
       simple error checking. If you get an error it will likely be here and it is because the cities database
       does not contain your location. Some options are given when an error arises.

       You are unlikely to want to call this function on its own. It is used inside the main function.
}
