% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.r
\name{distance.match.f}
\alias{distance.match.f}
\title{Lookup flying distances between locations (main DFO science locations)}
\usage{
distance.match.f(origin, destination)
}
\arguments{
\item{origin}{the origin city. It must be an exact match of the distance lookup table. No spaces and no
punctuation.}

\item{destination}{the destination for the activity. For a single activity in say Ottawa use
destination="Ottawa". If you put it in as a vector with "Ottawa" repeated as many times as the
origin vector, then just use the first column of the output.}
}
\description{
This looks up flying distances between origin and destination airports. Flying distances are
       generally taken as direct air distances between cities (https://canada-map.com/distance/). For
       smaller airports, where doubling back usually occurs then this is considered. For example Mont-Joli
       to St. John's requires that you fly southwest to Montreal before flying over Mont-Joli again
       northeast to get to St. John's. There we no effort to account for example that essentially straight
       flights would still have layovers at intermediate airports that could add km to the actual flying
       distance. The multiple possible routings can easily violate assumptions made in that kind of
       consideration.
}
\examples{
       # for the exact names in the distance table
       names(flying.distances)

       # An example use the TESA.course data and Ottawa as destination
       distance.match.f(TESA.course$Home, "Ottawa")

       # You will see that the last three values are NA and that is because neither Burlington nor
       # Copenhagen have entries in the distance table
}
\author{
Daniel Duplisea
}
