% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{ICES}
\alias{ICES}
\title{ICES}
\format{A data.frame or data.table or tibble}
\usage{
data(ICES)
}
\description{
A data.frame with a mock example of how a course program for ICES would need to be setup in order to calculate
the carbon footprint for training courses.
}
\details{
\itemize{
  \item activity.type the type of activity. You choose. e.g. online course, course, workshop, working group ...
  \item activity.name whatever you want to name the activity for the output figure
  \item origin the home city of the participant. Please try to choose the closest centre with >1000 people. For small places, e.g. Faroe Islands, you can use small centres
  \item destination the distination city for the activity
  \item origin.country the home country of the participant
  \item hotel.nights the number of nights the participant will sleep in a hotel. 0 if none
  \item meals the total number of meals the participant will eat in a restaurant. 0 if none
  \item bustrain.distance the total distance that a participant will take the bus or train. 0 if none
  \item car.distance the total distance that a participant will drive a car. 0 if none
  \item flying logical, if flying or not, 1 for yes 0 for no.
}
}
\keyword{datasets}
